+++ src/Translations_Updater/Init.php
@@ -26,6 +26,22 @@ class Init {
 	use Base;
 
 	/**
+	 * Holds calling class for EDD SL Updater.
+	 *
+	 * @var string
+	 */
+	private $caller;
+
+	/**
+	 * Constructor.
+	 *
+	 * @param string $caller Namespace of calling class.
+	 */
+	public function __construct( $caller = null ) {
+		$this->caller = $caller;
+	}
+
+	/**
 	 * Test for proper user capabilities.
 	 *
 	 * @return bool
@@ -69,30 +85,30 @@ class Init {
 	 * Load relevant action hooks for EDD Software Licensing.
 	 */
 	public function edd_run() {
-		add_action(
-			'post_edd_sl_plugin_updater_setup',
-			function ( $edd_config ) {
-				foreach ( $edd_config as $slug => $config ) {
-					if ( ! is_array( $config ) ) {
-						return false;
-					}
-					$config['type'] = 'plugin';
-					$config['slug'] = $slug;
-					$this->run( $config );
-				}
-			},
-			15,
-			1
-		);
-		add_action(
-			'post_edd_sl_theme_updater_setup',
-			function ( $config ) {
+		add_action( 'post_edd_sl_plugin_updater_setup', [ $this, 'parse_edd_config' ], 15, 1 );
+		add_action( 'post_edd_sl_theme_updater_setup', [ $this, 'parse_edd_config' ], 15, 1 );
+	}
+
+	/**
+	 * Parse passed config from EDD SL.
+	 *
+	 * @param array $config EDD SL config array.
+	 *
+	 * @return void
+	 */
+	public function parse_edd_config( $config ) {
+		$edd_sl_updater = 'EDD\Software_Licensing\Updater';
+		if ( $edd_sl_updater !== $this->caller ) {
+			if ( 'post_edd_sl_plugin_updater_setup' === current_filter() ) {
+				$config         = array_values( $config )[0];
+				$config['type'] = 'plugin';
+				$config['slug'] = $slug;
+			}
+			if ( 'post_edd_sl_theme_updater_setup' === current_filter() ) {
 				$config['type'] = 'theme';
 				$config['slug'] = $config['theme_slug'];
-				$this->run( $config );
-			},
-			15,
-			1
-		);
+			}
+		}
+		$this->run( $config );
 	}
 }