+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* update for possible universal EDD SL Updater plugin
 * switch to `site_transient_update_{plugins|themes}` filter
 * convert to composer dependency from [EDD Translations Updater](https://github.com/afragen/edd-translations-updater) and make more generic for any WordPress plugin or theme
 * support EDD Software Licensing `post_edd_sl_{plugin|theme}_updater_setup` action hooks